Compile for SDK 27 and rework gradle scripts
Upgrading to SDK 27 allows use of the newest gradle plugin. The gradle build files of the client and the iDPProxy have been reworked in order to provide the following flavour dimensions:
- app (cloud, cloudStaging, standalone)
- enrollment (autoEnrollmentDisabled, autoEnrollmentEnabled)
- type (BYOD, COPE)

Using product flavors it is easier to build different versions of the app in different build variants (release, debug). Including common configurations in the defaultConfig section eliminates some of the repetitions made.