Extract device CPU info correctly at enrollment for Android API 26
The Android agent reads and stores device CPU related data at the time
of enrollment. This data is not displayed, only retrieved and stored.
Up until recently the format of data when retrieved had remained
unchanged. With Android 26 the format has been updated. While this does
not cause the agent to break, the values retrieved are erroneous.

Furthermore when testing, one specific device; the OnePlus 5T was found
to have an entirely different format to that of regular Android. Due to
this the Android agent was prone to crashing when launched.

This commit addresses the incorrect retrieval of data of Android 26 and
also the app crash in the OnePlus 5T device.