make remote control feature mutual ssl compatible
Previously this operation logic was handled via the access tokens. Hence, in a situation where the device is enrolled under mutual-ssl, there won't be any access tokens to establish the socket connection as per the previous logic. With this change there is a new UUID introduced instead of the access token.